styling stuffz

+++ FSND-Virtual-Machine/vagrant/items_catalog/finalProject.py
@@ -33,18 +33,30 @@ session = DBSession()
 
 #HTML escape tags are like this: & becomes &amp, < becomees &lt, etc
 
-# # JSON API ENDPOINT FOR ITEMS
-# @app.route(''/category/<int:category_id>/item/<int:item_id>/JSON')
-# def itemJSON(category_id, item_id):
-#     category = session.query(Category).filter_by(id=category_id).one()
-#     items = session.query(Item).filter_by(category_id=category_id).all()
-#     return jsonify(Items=[i.serialize for i in items])
+# JSON API ENDPOINT FOR ALL ITEMS IN A CATEGORY
+@app.route('/category/<int:category_id>/items/JSON')
+def itemsJSON(category_id):
+    #category = session.query(Category).filter_by(id=category_id).one()
+    items = session.query(Item).filter_by(category_id=category_id).all()
+    return jsonify(items=[i.serialize for i in items])
 
-# JSON API ENDPOINT FOR CATEGORIES
+# JSON API ENDPOINT FOR A ITEM IN A CATEGORY
+@app.route('/item/<int:item_id>/JSON')
+def singleItemJSON(item_id):
+    items = session.query(Item).filter_by(id=item_id).all()
+    return jsonify(item=[i.serialize for i in items])
+
+# JSON API ENDPOINT FOR ALL CATEGORIES
+@app.route('/category/JSON')
+def categoriesJSON():
+    categories = session.query(Category).all()
+    return jsonify(categories=[c.serialize for c in categories])
+
+# JSON API ENDPOINT FOR ONE CATEGORY
 @app.route('/category/<int:category_id>/JSON')
 def categoryJSON(category_id):
-    categories = session.query(Category).filter_by(id=category_id).all()
-    return jsonify(Categories=[c.serialize for c in categories])
+    categories = session.query(Category).filter_by(id=category_id).all() # this returns only one object
+    return jsonify(category=[c.serialize for c in categories])
 
 
 # Show all categories
@@ -57,21 +69,6 @@ def showAllCategories():
         return render_template('categories.html', categories=allCategories)
     else:
         return render_template('public_categories.html', categories=allCategories)
-    # TODO: Render all Categories, not single category
-
-    # cat = session.query(Category)
-    # firstCategory = cat.first()
-    
-    # allItems = session.query(Item)
-    # itemCatId = allItems.filter_by(category_id=firstCategory.id)
-
-    # return render_template(
-    #   'categories.html',
-    #   blah={ 'name': 'blah', 'id': 1 },
-    #   category=firstCategory,
-    #   items=itemCatId
-    # ) # variable on LEFT is in the template, variable RIGHT is in the function
-
 
 # Add new category
 @app.route('/category/new/', methods=['GET','POST'])
@@ -81,7 +78,7 @@ def newCategory():
         return redirect('/login')
     if request.method == 'POST':
         print 'create new catalog'
-        newCategory = Category(name=request.form['categoryName'])
+        newCategory = Category(name=request.form['categoryName'], user_id=login_session['user_id'])
         session.add(newCategory)
         session.commit()
         flash("category has been added!")
@@ -91,10 +88,15 @@ def newCategory():
 
 # Edit category
 @app.route('/category/<int:category_id>/edit/', methods=['GET', 'POST'])
+
 def editCategory(category_id): #/category/1/edit
     if 'username' not in login_session:
         return redirect('/login')
     editedCategory = session.query(Category).filter_by(id=category_id).one()
+
+    if editedCategory.user_id != login_session['user_id']:
+        return redirect('/')
+
     if request.method == 'POST':
         if request.form['category_name']:
             editedCategory.name = request.form['category_name']
@@ -111,7 +113,12 @@ def editCategory(category_id): #/category/1/edit
 def deleteCategory(category_id):
     if 'username' not in login_session:
         return redirect('/login')
+
     categoryToDelete = session.query(Category).filter_by(id=category_id).one()
+
+    if categoryToDelete.user_id != login_session['user_id']:
+        return redirect('/')
+
     if request.method == 'POST':
         session.delete(categoryToDelete)
         session.commit()
@@ -134,9 +141,18 @@ def categoryPage(category_id):
 def addItem(category_id):
     if 'username' not in login_session:
         return redirect('/login')
+
+    category = session.query(Category).filter_by(id=category_id).one()
+
+    if not category:
+        return redirect('/')
+
+    if category.user_id != login_session['user_id']:
+        return redirect('/')
+    
     if request.method == 'POST':
         newItem = Item(name=request.form['itemName'], description=request.form['descriptionItem'],
-            category_id=category_id)
+            category_id=category.id, user_id=login_session['user_id'])
         session.add(newItem)
         session.commit()
         flash("New item %s has been added!" % (newItem.name))
@@ -146,14 +162,23 @@ def addItem(category_id):
 
 
 #Edit item
+#@login_required
+#@user_owns_post
 @app.route('/category/<int:category_id>/item/<int:item_id>/edit/', methods=['GET', 'POST'])
 def editItem(category_id, item_id):
     if 'username' not in login_session:
         return redirect('/login')
+    
     editedItem = session.query(Item).filter_by(id=item_id).one()
+
+    if editedItem.user_id != login_session['user_id']:
+        return redirect('/')
+
+    #editedDescription = sesssion.query(Item).filter_by(id=description_id).one()
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
+        #elif request.form['']
         session.add(editedItem)
         session.commit()
         flash("item has been edited!")
@@ -168,6 +193,10 @@ def deleteItem(category_id, item_id):
     if 'username' not in login_session:
         return redirect('/login')
     itemToDelete = session.query(Item).filter_by(id=item_id).one()
+
+    if itemToDelete.user_id != login_session['user_id']:
+        return redirect('/')
+
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
